Update models.py
Matches the New Schema with a few unknowns due to djangos capabilities. Unsure about adding the many to many tables due to Django being able to include ManyToMany Fields as well as OneToOne Fields. Possible that we won't need to create a separate class for this and that Django creates it itself but again unsure.